backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
ultralytics==8.1.0
easyocr==1.7.0
passporteye==2.2.0
pydantic==2.5.3
pydantic-settings==2.1.0